requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
Jinja2
requests
python-multipart
itsdangerous